Crate webview2_com

Modules

Microsoft

Structs

AcceleratorKeyPressedEventHandler

Implementation of ICoreWebView2AcceleratorKeyPressedEventHandler_Impl.

AcceleratorKeyPressedEventHandler_Impl

AddScriptToExecuteOnDocumentCreatedCompletedHandler

Implementation of ICoreWebView2AddScriptToExecuteOnDocumentCreatedCompletedHandler_Impl.

AddScriptToExecuteOnDocumentCreatedCompletedHandler_Impl

BasicAuthenticationRequestedEventHandler

Implementation of ICoreWebView2BasicAuthenticationRequestedEventHandler_Impl.

BasicAuthenticationRequestedEventHandler_Impl

BrowserExtensionEnableCompletedHandler

Implementation of ICoreWebView2BrowserExtensionEnableCompletedHandler_Impl.

BrowserExtensionEnableCompletedHandler_Impl

BrowserExtensionRemoveCompletedHandler

Implementation of ICoreWebView2BrowserExtensionRemoveCompletedHandler_Impl.

BrowserExtensionRemoveCompletedHandler_Impl

BrowserProcessExitedEventHandler

Implementation of ICoreWebView2BrowserProcessExitedEventHandler_Impl.

BrowserProcessExitedEventHandler_Impl

BytesReceivedChangedEventHandler

Implementation of ICoreWebView2BytesReceivedChangedEventHandler_Impl.

BytesReceivedChangedEventHandler_Impl

CallDevToolsProtocolMethodCompletedHandler

Implementation of ICoreWebView2CallDevToolsProtocolMethodCompletedHandler_Impl.

CallDevToolsProtocolMethodCompletedHandler_Impl

CapturePreviewCompletedHandler

Implementation of ICoreWebView2CapturePreviewCompletedHandler_Impl.

CapturePreviewCompletedHandler_Impl

ClearBrowsingDataCompletedHandler

Implementation of ICoreWebView2ClearBrowsingDataCompletedHandler_Impl.

ClearBrowsingDataCompletedHandler_Impl

ClearServerCertificateErrorActionsCompletedHandler

Implementation of ICoreWebView2ClearServerCertificateErrorActionsCompletedHandler_Impl.

ClearServerCertificateErrorActionsCompletedHandler_Impl

ClientCertificateRequestedEventHandler

Implementation of ICoreWebView2ClientCertificateRequestedEventHandler_Impl.

ClientCertificateRequestedEventHandler_Impl

CoTaskMemMut

Mutable guard object tied to the lifetime of the CoTaskMemPWSTR so that it is safe to dereference the PWSTR as long as both are still in scope.

CoTaskMemPWSTR

RAII holder for a PWSTR which is allocated with [Com::CoTaskMemAlloc] and freed with [Com::CoTaskMemFree] when dropped.

CoTaskMemRef

Constant guard object tied to the lifetime of the CoTaskMemPWSTR so that it is safe to dereference the PCWSTR as long as both are still in scope.

ContainsFullScreenElementChangedEventHandler

Implementation of ICoreWebView2ContainsFullScreenElementChangedEventHandler_Impl.

ContainsFullScreenElementChangedEventHandler_Impl

ContentLoadingEventHandler

Implementation of ICoreWebView2ContentLoadingEventHandler_Impl.

ContentLoadingEventHandler_Impl

ContextMenuRequestedEventHandler

Implementation of ICoreWebView2ContextMenuRequestedEventHandler_Impl.

ContextMenuRequestedEventHandler_Impl

CoreWebView2CustomSchemeRegistration

CoreWebView2CustomSchemeRegistration_Impl

CoreWebView2EnvironmentOptions

CoreWebView2EnvironmentOptions_Impl

CreateCoreWebView2CompositionControllerCompletedHandler

Implementation of ICoreWebView2CreateCoreWebView2CompositionControllerCompletedHandler_Impl.

CreateCoreWebView2CompositionControllerCompletedHandler_Impl

CreateCoreWebView2ControllerCompletedHandler

Implementation of ICoreWebView2CreateCoreWebView2ControllerCompletedHandler_Impl.

CreateCoreWebView2ControllerCompletedHandler_Impl

CreateCoreWebView2EnvironmentCompletedHandler

Implementation of ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler_Impl.

CreateCoreWebView2EnvironmentCompletedHandler_Impl

CursorChangedEventHandler

Implementation of ICoreWebView2CursorChangedEventHandler_Impl.

CursorChangedEventHandler_Impl

CustomItemSelectedEventHandler

Implementation of ICoreWebView2CustomItemSelectedEventHandler_Impl.

CustomItemSelectedEventHandler_Impl

DOMContentLoadedEventHandler

Implementation of ICoreWebView2DOMContentLoadedEventHandler_Impl.

DOMContentLoadedEventHandler_Impl

DevToolsProtocolEventReceivedEventHandler

Implementation of ICoreWebView2DevToolsProtocolEventReceivedEventHandler_Impl.

DevToolsProtocolEventReceivedEventHandler_Impl

DocumentTitleChangedEventHandler

Implementation of ICoreWebView2DocumentTitleChangedEventHandler_Impl.

DocumentTitleChangedEventHandler_Impl

DownloadStartingEventHandler

Implementation of ICoreWebView2DownloadStartingEventHandler_Impl.

DownloadStartingEventHandler_Impl

EstimatedEndTimeChangedEventHandler

Implementation of ICoreWebView2EstimatedEndTimeChangedEventHandler_Impl.

EstimatedEndTimeChangedEventHandler_Impl

ExecuteScriptCompletedHandler

Implementation of ICoreWebView2ExecuteScriptCompletedHandler_Impl.

ExecuteScriptCompletedHandler_Impl

ExecuteScriptWithResultCompletedHandler

Implementation of ICoreWebView2ExecuteScriptWithResultCompletedHandler_Impl.

ExecuteScriptWithResultCompletedHandler_Impl

FaviconChangedEventHandler

Implementation of ICoreWebView2FaviconChangedEventHandler_Impl.

FaviconChangedEventHandler_Impl

FocusChangedEventHandler

Implementation of ICoreWebView2FocusChangedEventHandler_Impl.

FocusChangedEventHandler_Impl

FrameContentLoadingEventHandler

Implementation of ICoreWebView2FrameContentLoadingEventHandler_Impl.

FrameContentLoadingEventHandler_Impl

FrameCreatedEventHandler

Implementation of ICoreWebView2FrameCreatedEventHandler_Impl.

FrameCreatedEventHandler_Impl

FrameDOMContentLoadedEventHandler

Implementation of ICoreWebView2FrameDOMContentLoadedEventHandler_Impl.

FrameDOMContentLoadedEventHandler_Impl

FrameDestroyedEventHandler

Implementation of ICoreWebView2FrameDestroyedEventHandler_Impl.

FrameDestroyedEventHandler_Impl

FrameNameChangedEventHandler

Implementation of ICoreWebView2FrameNameChangedEventHandler_Impl.

FrameNameChangedEventHandler_Impl

FrameNavigationCompletedEventHandler

Implementation of ICoreWebView2FrameNavigationCompletedEventHandler_Impl.

FrameNavigationCompletedEventHandler_Impl

FrameNavigationStartingEventHandler

Implementation of ICoreWebView2FrameNavigationStartingEventHandler_Impl.

FrameNavigationStartingEventHandler_Impl

FramePermissionRequestedEventHandler

Implementation of ICoreWebView2FramePermissionRequestedEventHandler_Impl.

FramePermissionRequestedEventHandler_Impl

FrameWebMessageReceivedEventHandler

Implementation of ICoreWebView2FrameWebMessageReceivedEventHandler_Impl.

FrameWebMessageReceivedEventHandler_Impl

GetCookiesCompletedHandler

Implementation of ICoreWebView2GetCookiesCompletedHandler_Impl.

GetCookiesCompletedHandler_Impl

GetFaviconCompletedHandler

Implementation of ICoreWebView2GetFaviconCompletedHandler_Impl.

GetFaviconCompletedHandler_Impl

GetNonDefaultPermissionSettingsCompletedHandler

Implementation of ICoreWebView2GetNonDefaultPermissionSettingsCompletedHandler_Impl.

GetNonDefaultPermissionSettingsCompletedHandler_Impl

GetProcessExtendedInfosCompletedHandler

Implementation of ICoreWebView2GetProcessExtendedInfosCompletedHandler_Impl.

GetProcessExtendedInfosCompletedHandler_Impl

HistoryChangedEventHandler

Implementation of ICoreWebView2HistoryChangedEventHandler_Impl.

HistoryChangedEventHandler_Impl

IFixedEnvironmentOptions4

This is an alternate declaration of the crate::Microsoft::Web::WebView2::Win32::ICoreWebView2EnvironmentOptions4 interface, which matches the parameters for SetCustomSchemeRegistrations. The windows crate mistakenly interprets the array of interface pointers as a pointer to an out-param, and it converts that into a Result<Option<ICoreWebView2CustomSchemeRegistration>>.

IsDefaultDownloadDialogOpenChangedEventHandler

Implementation of ICoreWebView2IsDefaultDownloadDialogOpenChangedEventHandler_Impl.

IsDefaultDownloadDialogOpenChangedEventHandler_Impl

IsDocumentPlayingAudioChangedEventHandler

Implementation of ICoreWebView2IsDocumentPlayingAudioChangedEventHandler_Impl.

IsDocumentPlayingAudioChangedEventHandler_Impl

IsMutedChangedEventHandler

Implementation of ICoreWebView2IsMutedChangedEventHandler_Impl.

IsMutedChangedEventHandler_Impl

LaunchingExternalUriSchemeEventHandler

Implementation of ICoreWebView2LaunchingExternalUriSchemeEventHandler_Impl.

LaunchingExternalUriSchemeEventHandler_Impl

MoveFocusRequestedEventHandler

Implementation of ICoreWebView2MoveFocusRequestedEventHandler_Impl.

MoveFocusRequestedEventHandler_Impl

NavigationCompletedEventHandler

Implementation of ICoreWebView2NavigationCompletedEventHandler_Impl.

NavigationCompletedEventHandler_Impl

NavigationStartingEventHandler

Implementation of ICoreWebView2NavigationStartingEventHandler_Impl.

NavigationStartingEventHandler_Impl

NewBrowserVersionAvailableEventHandler

Implementation of ICoreWebView2NewBrowserVersionAvailableEventHandler_Impl.

NewBrowserVersionAvailableEventHandler_Impl

NewWindowRequestedEventHandler

Implementation of ICoreWebView2NewWindowRequestedEventHandler_Impl.

NewWindowRequestedEventHandler_Impl

NonClientRegionChangedEventHandler

Implementation of ICoreWebView2NonClientRegionChangedEventHandler_Impl.

NonClientRegionChangedEventHandler_Impl

PermissionRequestedEventHandler

Implementation of ICoreWebView2PermissionRequestedEventHandler_Impl.

PermissionRequestedEventHandler_Impl

PrintCompletedHandler

Implementation of ICoreWebView2PrintCompletedHandler_Impl.

PrintCompletedHandler_Impl

PrintToPdfCompletedHandler

Implementation of ICoreWebView2PrintToPdfCompletedHandler_Impl.

PrintToPdfCompletedHandler_Impl

PrintToPdfStreamCompletedHandler

Implementation of ICoreWebView2PrintToPdfStreamCompletedHandler_Impl.

PrintToPdfStreamCompletedHandler_Impl

ProcessFailedEventHandler

Implementation of ICoreWebView2ProcessFailedEventHandler_Impl.

ProcessFailedEventHandler_Impl

ProcessInfosChangedEventHandler

Implementation of ICoreWebView2ProcessInfosChangedEventHandler_Impl.

ProcessInfosChangedEventHandler_Impl

ProfileAddBrowserExtensionCompletedHandler

Implementation of ICoreWebView2ProfileAddBrowserExtensionCompletedHandler_Impl.

ProfileAddBrowserExtensionCompletedHandler_Impl

ProfileDeletedEventHandler

Implementation of ICoreWebView2ProfileDeletedEventHandler_Impl.

ProfileDeletedEventHandler_Impl

ProfileGetBrowserExtensionsCompletedHandler

Implementation of ICoreWebView2ProfileGetBrowserExtensionsCompletedHandler_Impl.

ProfileGetBrowserExtensionsCompletedHandler_Impl

RasterizationScaleChangedEventHandler

Implementation of ICoreWebView2RasterizationScaleChangedEventHandler_Impl.

RasterizationScaleChangedEventHandler_Impl

ScriptDialogOpeningEventHandler

Implementation of ICoreWebView2ScriptDialogOpeningEventHandler_Impl.

ScriptDialogOpeningEventHandler_Impl

ServerCertificateErrorDetectedEventHandler

Implementation of ICoreWebView2ServerCertificateErrorDetectedEventHandler_Impl.

ServerCertificateErrorDetectedEventHandler_Impl

SetPermissionStateCompletedHandler

Implementation of ICoreWebView2SetPermissionStateCompletedHandler_Impl.

SetPermissionStateCompletedHandler_Impl

SourceChangedEventHandler

Implementation of ICoreWebView2SourceChangedEventHandler_Impl.

SourceChangedEventHandler_Impl

StateChangedEventHandler

Implementation of ICoreWebView2StateChangedEventHandler_Impl.

StateChangedEventHandler_Impl

StatusBarTextChangedEventHandler

Implementation of ICoreWebView2StatusBarTextChangedEventHandler_Impl.

StatusBarTextChangedEventHandler_Impl

TrySuspendCompletedHandler

Implementation of ICoreWebView2TrySuspendCompletedHandler_Impl.

TrySuspendCompletedHandler_Impl

WebMessageReceivedEventHandler

Implementation of ICoreWebView2WebMessageReceivedEventHandler_Impl.

WebMessageReceivedEventHandler_Impl

WebResourceRequestedEventHandler

Implementation of ICoreWebView2WebResourceRequestedEventHandler_Impl.

WebResourceRequestedEventHandler_Impl

WebResourceResponseReceivedEventHandler

Implementation of ICoreWebView2WebResourceResponseReceivedEventHandler_Impl.

WebResourceResponseReceivedEventHandler_Impl

WebResourceResponseViewGetContentCompletedHandler

Implementation of ICoreWebView2WebResourceResponseViewGetContentCompletedHandler_Impl.

WebResourceResponseViewGetContentCompletedHandler_Impl

WindowCloseRequestedEventHandler

Implementation of ICoreWebView2WindowCloseRequestedEventHandler_Impl.

WindowCloseRequestedEventHandler_Impl

ZoomFactorChangedEventHandler

Implementation of ICoreWebView2ZoomFactorChangedEventHandler_Impl.

ZoomFactorChangedEventHandler_Impl

Enums

Error

Traits

ClosureArg

IFixedEnvironmentOptions4_Impl

InvokeArg

Functions

pwstr_from_str

Allocate a PWSTR with [Com::CoTaskMemAlloc] and copy a &str into it.

string_from_pcwstr

Copy a PCWSTR from an input param to a String.

take_pwstr

Copy a PWSTR allocated with [Com::CoTaskMemAlloc] from an input param to a String and free the original buffer with [Com::CoTaskMemFree].

wait_with_pump

The WebView2 threading model runs everything on the UI thread, including callbacks which it triggers with PostMessage, and we’re using this here because it’s waiting for some async operations in WebView2 to finish before starting the main message loop. As long as there are no pending results in rx, it will pump Window messages and check for a result after each message is dispatched.

Type Aliases

CompletedClosure

Generic closure signature for [completed_callback].

EventClosure

Generic closure signature for [event_callback].

Result