Struct webview2_com::CoreWebView2EnvironmentOptions

pub struct CoreWebView2EnvironmentOptions { /* private fields */ }

Implementations

impl CoreWebView2EnvironmentOptions

pub unsafe fn additional_browser_arguments(&self) -> &str

Equivalent to ICoreWebView2EnvironmentOptions::AdditionalBrowserArguments without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<String> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_additional_browser_arguments(&self, value: String)

Equivalent to ICoreWebView2EnvironmentOptions::SetAdditionalBrowserArguments without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<String> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn language(&self) -> &str

Equivalent to ICoreWebView2EnvironmentOptions::Language without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<String> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_language(&self, value: String)

Equivalent to ICoreWebView2EnvironmentOptions::SetLanguage without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<String> as the COM method, but it does not dereference a pointer for the value parameter. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn target_compatible_browser_version(&self) -> &str

Equivalent to ICoreWebView2EnvironmentOptions::TargetCompatibleBrowserVersion without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<String> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_target_compatible_browser_version(&self, value: String)

Equivalent to ICoreWebView2EnvironmentOptions::SetTargetCompatibleBrowserVersion without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<String> as the COM method, but it does not dereference a pointer for the value parameter. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn allow_single_sign_on_using_os_primary_account(&self) -> bool

Equivalent to ICoreWebView2EnvironmentOptions::AllowSingleSignOnUsingOSPrimaryAccount without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<bool> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_allow_single_sign_on_using_os_primary_account( &self, value: bool, )

Equivalent to ICoreWebView2EnvironmentOptions::SetAllowSingleSignOnUsingOSPrimaryAccount without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<bool> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn exclusive_user_data_folder_access(&self) -> bool

Equivalent to ICoreWebView2EnvironmentOptions2::ExclusiveUserDataFolderAccess without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<bool> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_exclusive_user_data_folder_access(&self, value: bool)

Equivalent to ICoreWebView2EnvironmentOptions2::SetExclusiveUserDataFolderAccess without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<bool> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn is_custom_crash_reporting_enabled(&self) -> bool

Equivalent to ICoreWebView2EnvironmentOptions3::IsCustomCrashReportingEnabled without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<bool> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_is_custom_crash_reporting_enabled(&self, value: bool)

Equivalent to ICoreWebView2EnvironmentOptions3::SetIsCustomCrashReportingEnabled without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<bool> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn scheme_registrations( &self, ) -> Vec<Option<ICoreWebView2CustomSchemeRegistration>>

Equivalent to IFixedEnvironmentOptions4::GetCustomSchemeRegistrations without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<Vec<>> as the COM method, but it does not allocate or write to a mutable pointer for the result.

pub unsafe fn set_scheme_registrations( &self, value: Vec<Option<ICoreWebView2CustomSchemeRegistration>>, )

Equivalent to IFixedEnvironmentOptions4::SetCustomSchemeRegistrations without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<Vec<>> as the COM method. It takes an immutable reference to self for consistency with the other public set methods, however the COM method implementation does not actually reuse it.

pub unsafe fn enable_tracking_prevention(&self) -> bool

Equivalent to ICoreWebView2EnvironmentOptions5::EnableTrackingPrevention without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<bool> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_enable_tracking_prevention(&self, value: bool)

Equivalent to ICoreWebView2EnvironmentOptions5::SetEnableTrackingPrevention without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<bool> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn are_browser_extensions_enabled(&self) -> bool

Equivalent to ICoreWebView2EnvironmentOptions6::AreBrowserExtensionsEnabled without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<bool> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_are_browser_extensions_enabled(&self, value: bool)

Equivalent to ICoreWebView2EnvironmentOptions6::SetAreBrowserExtensionsEnabled without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<bool> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn channel_search_kind(&self) -> COREWEBVIEW2_CHANNEL_SEARCH_KIND

Equivalent to ICoreWebView2EnvironmentOptions7::ChannelSearchKind without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<COREWEBVIEW2_CHANNEL_SEARCH_KIND> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_channel_search_kind( &self, value: COREWEBVIEW2_CHANNEL_SEARCH_KIND, )

Equivalent to ICoreWebView2EnvironmentOptions7::SetChannelSearchKind without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<COREWEBVIEW2_CHANNEL_SEARCH_KIND> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn release_channels(&self) -> COREWEBVIEW2_RELEASE_CHANNELS

Equivalent to ICoreWebView2EnvironmentOptions7::ReleaseChannels without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<COREWEBVIEW2_RELEASE_CHANNELS> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_release_channels(&self, value: COREWEBVIEW2_RELEASE_CHANNELS)

Equivalent to ICoreWebView2EnvironmentOptions7::SetReleaseChannels without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<COREWEBVIEW2_RELEASE_CHANNELS> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.

pub unsafe fn scroll_bar_style(&self) -> COREWEBVIEW2_SCROLLBAR_STYLE

Equivalent to ICoreWebView2EnvironmentOptions8::ScrollBarStyle without extra memory copies or type conversions.

Safety

This method reads from the same UnsafeCell<COREWEBVIEW2_SCROLLBAR_STYLE> as the COM method, but it does not write to a mutable pointer for the result.

pub unsafe fn set_scroll_bar_style(&self, value: COREWEBVIEW2_SCROLLBAR_STYLE)

Equivalent to ICoreWebView2EnvironmentOptions8::SetScrollBarStyle without extra memory copies or type conversions.

Safety

This method writes to the same UnsafeCell<COREWEBVIEW2_SCROLLBAR_STYLE> as the COM method. It takes an immutable reference to self so that it can be reused in the COM method.